Change message from "warning" to "info".
It is a perfectly normal behaviour in cases where you have both Direct
Readout (DR) and RARS (e.g.. EARS) scenes. Often the DR scene will come
first and it may be convenient then to skip processing the "same" scene
from the RARS stream. This is assuming they cover the same geographical
area approximately. This can be controlled via the "collection_area_id"
in the posttroll message being input to the AAPP-runner

Signed-off-by: Adam.Dybbroe <a000680@c21856.ad.smhi.se>